+++ CHANGELOG.md
@@ -20,6 +20,7 @@ NOTE: Worlds that used Refined Storage 1.5.x are fully compatible with Refined S
 - Fixed a crash when breaking an Ender IO conduit with the Destructor (raoulvdberge)
 - Fixed bug where storage disks in Portable Grids could be moved into themselves (raoulvdberge)
 - Fixed the Crafter crashing when opening it while connected to a Primal Tech Grill or Kiln (raoulvdberge)
+- Fixed bug where Crafting Upgrade on Interface kept too many items in stock (raoulvdberge)
 - Prevent accidental Grid scrollbar click after clicking JEI recipe transfer button (raoulvdberge)
 - Added a missing config option for Crafter Manager energy usage (raoulvdberge)
 - If an Interface is configured to expose the entire network storage (by configuring no export slots), it will no longer expose the entire RS storage, due to performance issues (raoulvdberge)